# Optimize DuckDB Arrow Integration: 2.3x Performance Improvement