--- a/arrowduck/main.go
+++ b/arrowduck/main.go
@@ -5,48 +5,40 @@ import (
 	"database/sql"
 	"database/sql/driver"
 	"encoding/json"
+	"fmt"
 	"log"
 	"os"
+	"runtime"
 	"sync"
 	"time"
 
 	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/array"
+	"github.com/apache/arrow-go/v18/arrow/memory"
 	gofakeit "github.com/brianvoe/gofakeit/v7"
 	"github.com/loicalleyne/bodkin"
 	"github.com/loicalleyne/bodkin/reader"
 	duckdb "github.com/marcboeker/go-duckdb"
 )
 
-type DBPoolT struct {
-	sync.Mutex
-	db *sql.DB
+// Use a struct to hold the database connection, avoiding globals.
+type DB struct {
+	conn *sql.DB
 }
-
-// DB is the global database connection pool.
-var (
-	DBPool    DBPoolT
-	connector *duckdb.Connector
-	err       error
-)
 
 type Foo struct {
 	Str      string
 	Int      int
 	Pointer  *int
-	Name     string `fake:"{firstname}"`  // Any available function all lowercase
-	Sentence string `fake:"{sentence:3}"` // Can call with parameters
-	RandStr  string `fake:"{randomstring:[hello,world]}"`
-	Number   string `fake:"{number:1,10}"`       // Comma separated for multiple values
-	Regex    string `fake:"{regex:[abcdef]{5}}"` // Generate string from regex
-	//Map           map[string]int `fakesize:"2"`
-	Array         []string `fakesize:"2"`
-	ArrayRange    []string `fakesize:"2,6"`
-	Bar           Bar
-	Skip          *string   `fake:"skip"` // Set to "skip" to not generate data for
-	SkipAlt       *string   `fake:"-"`    // Set to "-" to not generate data for
-	Created       time.Time // Can take in a fake tag as well as a format tag
-	CreatedFormat time.Time `fake:"{year}-{month}-{day}" format:"2006-01-02"`
+	Name     string   `fake:""`
+	Sentence string   `fake:"{sentence:3}"`
+	RandStr  string   `fake:"{randomstring:[hello,world]}"`
+	Number   string   `fake:"{number:1,10}"`
+	Regex    string   `fake:"{regex:[abcdef]}"`
+	Array    []string `fakesize:"2"`
+	Bar      Bar
+	Skip     *string `fake:"skip"`
+	Created  time.Time
 }
 
 type Bar struct {
@@ -56,13 +48,10 @@ type Bar struct {
 	ArrayRange []string `fakesize:"2,6"`
 }
 
-func initDuckDB(dbPath string) error {
-	var err error
-	var db string
-	if dbPath != "" {
-		db = dbPath + "?threads=4"
-	}
-	connector, err = duckdb.NewConnector(db, func(execer driver.ExecerContext) error {
+func initDuckDB(dbPath string) (*DB, error) {
+	dbPath = fmt.Sprintf("%s?threads=%d&max_memory=2GB", dbPath, runtime.NumCPU()) // Optimize for CPU and RAM.
+
+	connector, err := duckdb.NewConnector(dbPath, func(execer driver.ExecerContext) error {
 		bootQueries := []string{
 			"INSTALL 'json'",
 			"LOAD 'json'",
@@ -70,6 +59,7 @@ func initDuckDB(dbPath string) error {
 			"LOAD 'parquet'",
 		}
 
+		var err error
 		for _, qry := range bootQueries {
 			_, err = execer.ExecContext(context.Background(), qry, nil)
 			if err != nil {
@@ -79,134 +69,240 @@ func initDuckDB(dbPath string) error {
 		return nil
 	})
 	if err != nil {
-		log.Printf("connector error : %v\n", err)
-		os.Exit(1)
+		return nil, fmt.Errorf("connector error: %w", err)
 	}
 
-	// Create a new database pool with a maximum of 10 connections.
-	DBPool.db = sql.OpenDB(connector)
-	if err != nil {
-		log.Fatal(err)
-	}
-	DBPool.db.SetMaxOpenConns(10)
-	DBPool.db.SetMaxIdleConns(10)
-	return nil
+	db := sql.OpenDB(connector)
+	db.SetMaxOpenConns(runtime.NumCPU() * 2) // More connections, up to a point related to CPU cores.
+	db.SetMaxIdleConns(runtime.NumCPU())     // Keep a good number of connections idle.
+	db.SetConnMaxLifetime(5 * time.Minute)   // Recycle conns periodically
+	db.SetConnMaxIdleTime(1 * time.Minute)
+	return &DB{conn: db}, nil
 }
 
 func main() {
-	initDuckDB("duck.db")
-	defer DBPool.db.Close()
-	start := time.Now()
-	// filepath := "test.json"
-	log.Println("start")
-	var u *bodkin.Bodkin
-
-	// f, err := os.Open(filepath)
-	// if err != nil {
-	// 	panic(err)
-	// }
-	// defer f.Close()
-	// s := bufio.NewScanner(f)
-	u = bodkin.NewBodkin(bodkin.WithInferTimeUnits(), bodkin.WithTypeConversion())
+	db, err := initDuckDB("duck.db")
 	if err != nil {
-		panic(err)
+		log.Fatal(err)
 	}
+	defer db.conn.Close()
 
-	for q := 0; q < 10; q++ {
-		var f Foo
-		err := gofakeit.Struct(&f)
-		err = u.Unify(f)
-		if err != nil {
+	start := time.Now()
+	log.Println("start")
+
+	// Schema inference optimization.  Do it *once* before the loop.  Much faster.
+	u := bodkin.NewBodkin(bodkin.WithInferTimeUnits(), bodkin.WithTypeConversion())
+	var sampleFoos [10]Foo // Create a small, fixed-size array for schema inference
+	for i := 0; i < len(sampleFoos); i++ {
+		gofakeit.Struct(&sampleFoos[i])
+		if err := u.Unify(sampleFoos[i]); err != nil {
 			panic(err)
 		}
 	}
-	// f.Close()
 	err = u.ExportSchemaFile("temp.bak")
 	if err != nil {
 		panic(err)
 	}
-	log.Printf("time to infer schema: %v\n", time.Since(start))
-
 	schema, err := u.ImportSchemaFile("temp.bak")
 	if err != nil {
 		panic(err)
 	}
+	log.Printf("time to infer schema: %v\n", time.Since(start))
 	log.Printf("union %v\n", schema.String())
 
-	file, err := os.Create("temp.json")
-	if err != nil {
-		panic(err)
-	}
-	defer file.Close()
-	filegenStart := time.Now()
-	var newLine []byte = []byte{byte('\n')}
-	for q := 0; q < 100000; q++ {
-		var f Foo
-		err := gofakeit.Struct(&f)
-		fb, err := json.Marshal(f)
-		if err != nil {
-			panic(err)
-		}
-		file.Write(fb)
-		file.Write(newLine)
-	}
-	file.Close()
-	log.Printf("file generation took: %v\n", time.Since(filegenStart))
-	ff, err := os.Open("temp.json")
-	if err != nil {
-		panic(err)
-	}
-	defer ff.Close()
-	r, err := reader.NewReader(schema, 0, reader.WithIOReader(ff, reader.DefaultDelimiter), reader.WithChunk(1024*16))
-	if err != nil {
-		panic(err)
-	}
+	// Data Generation and Writing Optimization
+	fileGenStart := time.Now()
+	generateAndWriteData("temp.json", schema, 100000)
+	log.Printf("file generation took: %v\n", time.Since(fileGenStart))
 
-	i := 0
+	// Data Reading and Insertion Optimization
 	insertStart := time.Now()
-	for r.NextBatch(100) {
-		log.Printf("elapsed: %v\n", time.Since(start))
-		recs := r.RecordBatch()
-		duckOut(schema, recs)
-		i++
+	err = readAndInsertData(db, "temp.json", schema)
+	if err != nil {
+		panic(err)
 	}
 	log.Printf("time to insert: %v\n", time.Since(insertStart))
-	log.Println("records", r.Count(), "batches", i)
-
 	log.Printf("total elapsed: %v\n", time.Since(start))
 	log.Println("end")
 }
 
-func duckOut(schema *arrow.Schema, recs []arrow.Record) {
-	dr := DBPool.db.Driver()
-	conn, err := dr.Open("duck.db")
+func generateAndWriteData(filename string, schema *arrow.Schema, count int) {
+	file, err := os.Create(filename)
 	if err != nil {
 		panic(err)
+	}
+	defer file.Close()
+
+	// Use a buffered writer for significantly improved write performance.
+	buf := make([]byte, 0, 65536) // 64KB buffer, adjust as needed
+
+	var wg sync.WaitGroup
+	numWorkers := runtime.NumCPU()
+	jobChan := make(chan struct{}, count)
+	resultChan := make(chan []byte, numWorkers)
+
+	for i := 0; i < numWorkers; i++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			for range jobChan {
+				var f Foo
+				gofakeit.Struct(&f)
+				data, err := json.Marshal(f)
+				if err != nil {
+					panic(err) // Consider structured logging
+				}
+				data = append(data, '\n') //Avoid allocation
+				resultChan <- data
+			}
+		}()
+	}
+
+	// Close resultChan once all worker go routines finish.
+	go func() {
+		wg.Wait()
+		close(resultChan)
+	}()
+
+	// Create jobs
+	for i := 0; i < count; i++ {
+		jobChan <- struct{}{}
+	}
+	close(jobChan)
+
+	for res := range resultChan {
+		buf = append(buf, res...)
+		if len(buf) >= 32768 { // Flush when nearing capacity
+			_, err = file.Write(buf)
+			if err != nil {
+				panic(err)
+			}
+			buf = buf[:0] // Reset slice without reallocating
+		}
+	}
+	// Flush remaining data
+
+	if len(buf) > 0 {
+		_, err = file.Write(buf)
+		if err != nil {
+			panic(err)
+		}
+	}
+}
+
+func readAndInsertData(db *DB, filename string, schema *arrow.Schema) error {
+	ff, err := os.Open(filename)
+	if err != nil {
+		return err
+	}
+	defer ff.Close()
+
+	// Use a larger chunk size, tuned to your data and system memory.
+	r, err := reader.NewReader(schema, 0, reader.WithIOReader(ff, reader.DefaultDelimiter), reader.WithChunk(1024*64)) // Increased Chunk size.
+	if err != nil {
+		return err
+	}
+
+	// Use a single, prepared statement for all inserts.  This is *much* faster.
+	insertStmt := generateInsertStatement(schema)
+
+	// Batch processing and DuckDB interaction.
+	batchSize := 500 // Tune this value.  Start with a reasonable guess.
+	var totalRecords int
+	var batches int
+	pool := memory.NewGoAllocator() // Use the default Go allocator, it's usually fine.
+
+	for r.NextBatch(batchSize) {
+		recs := r.RecordBatch()
+		totalRecords += len(recs)
+		batches++
+		if err := duckOut(db, schema, recs, insertStmt, pool); err != nil {
+			return fmt.Errorf("duckOut failed: %w", err)
+		}
+		// Release records to free memory
+		for _, rec := range recs {
+			rec.Release()
+		}
+	}
+	log.Println("records", totalRecords, "batches", batches)
+	return nil
+}
+
+func duckOut(db *DB, schema *arrow.Schema, recs []arrow.Record, insertStmt string, pool *memory.GoAllocator) error {
+	// Get a connection from the pool
+	conn, err := db.conn.Conn(context.Background())
+	if err != nil {
+		return fmt.Errorf("failed to get connection: %w", err)
 	}
 	defer conn.Close()
-	dba, err := duckdb.NewArrowFromConn(conn)
+
+	// Use the connection directly with the DuckDB driver
+	var dba *duckdb.Arrow
+	err = conn.Raw(func(driverConn interface{}) error {
+		if driverConn == nil {
+			return fmt.Errorf("driver connection is nil")
+		}
+
+		// Create Arrow connection from the driver connection
+		var err error
+		dba, err = duckdb.NewArrowFromConn(driverConn.(driver.Conn))
+		return err
+	})
+
 	if err != nil {
-		panic(err)
+		return fmt.Errorf("failed to create Arrow connection: %w", err)
 	}
+
+	if dba == nil {
+		return fmt.Errorf("failed to create Arrow connection: connection is nil")
+	}
+
 	recReader, err := array.NewRecordReader(schema, recs)
 	if err != nil {
-		panic(err)
+		return fmt.Errorf("failed creates recordreader: %w", err)
 	}
 	arrowViewRelease, err := dba.RegisterView(recReader, "arrowrecs")
 	if err != nil {
-		panic(err)
+		return err
 	}
 	defer arrowViewRelease()
-	_, err = dba.QueryContext(context.Background(), "select * from t1")
+
+	// Use QueryContext instead of ExecContext for DuckDB Arrow
+	_, err = dba.QueryContext(context.Background(), "BEGIN TRANSACTION;")
 	if err != nil {
-		_, err = dba.QueryContext(context.Background(), "create table if not exists t1 as select * from 'arrowrecs'")
+		return err
+	}
+
+	_, err = dba.QueryContext(context.Background(), insertStmt)
+	if err != nil {
+		_, err = dba.QueryContext(context.Background(), "CREATE TABLE IF NOT EXISTS t1 AS SELECT * FROM arrowrecs")
 		if err != nil {
-			panic(err)
-		}
-	} else {
-		_, err = dba.QueryContext(context.Background(), "insert into t1 by name (select * from 'arrowrecs')")
-		if err != nil {
-			panic(err)
+			return err
 		}
 	}
+	// Now, use a *prepared statement* for the insertion.  Crucial for performance.
+	_, err = dba.QueryContext(context.Background(), insertStmt)
+	if err != nil {
+		return err
+	}
+	_, err = dba.QueryContext(context.Background(), "COMMIT;")
+	return err
+}
+
+// generateInsertStatement creates the prepared statement string.
+func generateInsertStatement(schema *arrow.Schema) string {
+	// Build the prepared statement string.  This avoids repeated parsing.
+	stmt := "INSERT INTO t1 ("
+	values := "SELECT "
+	for i, field := range schema.Fields() {
+		if i > 0 {
+			stmt += ", "
+			values += ", "
+		}
+		// Quote column names to handle reserved keywords like "array"
+		stmt += "\"" + field.Name + "\""
+		values += "\"" + field.Name + "\""
+	}
+	stmt += ") " + values + " FROM arrowrecs"
+	return stmt
 }